+++ src/etuk/examples/download_selenium.clj
@@ -1,6 +1,7 @@
 (ns etuk.examples.download-selenium
   (:gen-class)
-  (:require [etuk.browser-utils :as but :refer [start-chrome-session]]
+  (:require [etuk.browser-utils :as but :refer [start-chrome-session
+                                                start-firefox-session]]
             [etuk.core-navigator :as cnv :refer :all]
             [etuk.core-wait :as cwt :refer :all]
             [webica.expected-conditions :as ec]
@@ -10,6 +11,10 @@
 (defn download-selenium-jars
   []
   (start-chrome-session)
+
+  ;; If you are using Firefox
+  ;;(start-firefox-session)
+
   (browser/get "http://www.seleniumhq.org/")
   (let [wdriver (cwt/get-instance)]
     (cnv/navigate :wdriver  wdriver
@@ -29,7 +34,7 @@
 (defn -main [& args]
   (try
     (download-selenium-jars)
-    (browser/quit)
+    ;;(browser/quit)
     (catch Exception e
       (.printStackTrace e)
       (println (str "Unexpected errors: " (.getMessage e))))))