+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## [0.1.2-SNAPSHOT] - 2017-06-14
+
+- Update dependencies to latest versions
+- Update Selenium to 3.4.0 (latest version)
+- Minor code update for Firefox browser
+
 ### Changed
 
 ## [0.1.1] - 2017-03-31
@@ -11,5 +17,6 @@ All notable changes to this project will be documented in this file. This change
 
 - Initial release
 
-[Unreleased]: https://github.com/agilecreativity/tuktuk/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/agilecreativity/tuktuk/compare/0.1.2-SNAPSHOT...HEAD
+[0.1.2-SNAPSHOT]: https://github.com/agilecreativity/tuktuk/compare/0.1.1...0.1.2-SNAPSHOT
 [0.1.1]: https://github.com/agilecreativity/tuktuk/compare/0.1.0...0.1.1